WIP: Allow users to specify bind IP for sessions.
When setting up implicit connections, EIPScanner was binding to
INADDR_ANY port 2222 to communicate with devices. This meant that only
one device at a time could be used. This code allows the user to specify
a session IP address that will be bound for implicit communications,
allowing a machine with multiple interfaces to connect to multiple
Ethernet/IP devices.